requirements.txt
datasets
sentence-transformers
faiss-cpu
requests
